Enhance UI and Accessibility in Todo App
In this update, we've made several improvements to the todo list application to enhance user experience and accessibility. First, a screenshot of the todo list app is now included in the README.md to give new users a visual understanding of the app's UI. This should make it easier for potential users to grasp the app's functionality at a glance.

We've adjusted the viewport settings in `index.html` for a better mobile experience, preventing users from scaling the page which ensures consistency in how the content is displayed across devices. The theme-color meta tag has also been updated to a lighter shade, improving readability and cohesiveness with the app's overall design.

In both the `App.css` and `App.scss` files, styles were refined to allow scrolling within the app container, addressing overflow issues that could hinder user interaction. Additionally, the footer's padding and font size were adjusted for better visibility and positioning across different devices.

Significant changes were made to the `AddTodo` component and its styling, improving the layout and interaction feedback. We wrapped the add-todo form within a new `add-todo-container` class, enhancing the structural organization, and introduced a hint text that appears when the input field is not empty, guiding users to press Enter to add their todo item. This small textual hint aids in understanding the app's functionality without cluttering the interface.

Furthermore, enhancements to the `TodoList` CSS include better visibility for delete buttons on hover and a more intuitive representation of the interactive checklist, improving both the usability and aesthetic appeal of the list items.

We also reversed the order todos are displayed by modifying the `TodoList` component to list the newest items first, making it easier for users to track their most recent tasks.

Lastly, smooth transitions for all elements were introduced to soften the visual changes and interactions within the app, making the user experience more pleasant and less jarring.

These changes collectively aim to refine the app's interface, making it more intuitive, accessible, and visually appealing, ultimately encouraging user engagement and improving usability.